feat: add not available status to partner offer notifications
