+++ src/Components/Notifications/ExpiresInTimer.tsx
@@ -1,61 +1,53 @@
 import StopwatchIcon from "@artsy/icons/StopwatchIcon"
 import { Flex, Text } from "@artsy/palette"
+import { formattedTimeLeft } from "Components/Notifications/util"
 import { useTimer } from "Utils/Hooks/useTimer"
 import { NotificationItem_item$data } from "__generated__/NotificationItem_item.graphql"
+import { FC } from "react"
 
-export const ExpiresInTimer: React.FC<{ expiresAt: string }> = ({
-  expiresAt,
-}) => {
+interface ExpiresInTimerProps {
+  item: NotificationItem_item$data
+}
+
+const WatchIcon: FC<{ fill?: string }> = ({ fill = "red100" }) => {
+  return <StopwatchIcon fill={fill} height={15} width={15} mr="2px" ml="-2px" />
+}
+
+export const ExpiresInTimer: FC<ExpiresInTimerProps> = ({ item }) => {
+  const expiresAt = item?.item?.expiresAt ?? ""
+  const available = item?.item?.available ?? false
+
   const { hasEnded, time } = useTimer(expiresAt)
 
-  if (hasEnded) {
+  if (!available) {
     return (
-      <Text variant="xs" color="red100">
-        Expired
-      </Text>
+      <Flex flexDirection="row" alignItems="center">
+        <WatchIcon />
+
+        <Text variant="xs" color="red100">
+          No longer available
+        </Text>
+      </Flex>
     )
   }
 
-  const formattedTimeLeft = (time: {
-    days: string
-    hours: string
-    minutes: string
-    seconds: string
-  }) => {
-    const parsedDays = parseInt(time.days, 10)
-    const parsedHours = parseInt(time.hours, 10)
-    const parsedMinutes = parseInt(time.minutes, 10)
-    const parsedSeconds = parseInt(time.seconds, 10)
+  if (hasEnded) {
+    return (
+      <Flex flexDirection="row" alignItems="center">
+        <WatchIcon />
 
-    let textColor = "blue100"
-    let copy
-
-    if (parsedDays >= 1) {
-      copy = `${parsedDays} days`
-    } else if (parsedDays < 1 && parsedHours >= 1) {
-      copy = `${parsedHours} hours`
-    } else if (parsedHours < 1 && parsedMinutes >= 1) {
-      copy = `${parsedMinutes} minutes`
-      textColor = "red100"
-    } else if (parsedMinutes < 1) {
-      copy = `${parsedSeconds} seconds`
-      textColor = "red100"
-    }
-
-    return { timerCopy: copy, textColor }
+        <Text variant="xs" color="red100">
+          Expired
+        </Text>
+      </Flex>
+    )
   }
 
   const { timerCopy, textColor } = formattedTimeLeft(time)
 
   return (
     <Flex flexDirection="row" alignItems="center">
-      <StopwatchIcon
-        fill={textColor}
-        height={15}
-        width={15}
-        mr="2px"
-        ml="-2px"
-      />
+      <WatchIcon fill={textColor} />
 
       <Text variant="xs" color={textColor}>
         Expires in {timerCopy}
@@ -68,8 +60,6 @@ export const shouldDisplayExpiresInTimer = (
   item: NotificationItem_item$data
 ) => {
   return (
-    item.notificationType === "PARTNER_OFFER_CREATED" &&
-    item.item &&
-    item.item.expiresAt
+    item.notificationType === "PARTNER_OFFER_CREATED" && item?.item?.expiresAt
   )
 }
